SG-56: Full logs widget
This adds the widget that display the full logs of the usage of the
garage door. This makes it easier to track history for various reasons.

HTTPS is also now disabled in frontend and monitoring backendbecause
of the latency from the backend server. HTTPS to the
garage-control-server remains the same.